refactor: relocate config.py -> pdft_benchmarks.presets
Pure file move + import-path rewrite across the 5 runner scripts and
harness.py + tests. Public API (Preset, get_preset, _DATASETS,
_BASE_PRESETS, PRESETS_*) unchanged.
