+++ run_quickdraw.py
@@ -36,7 +36,7 @@ from baselines import (
     global_dct_compress,
     global_fft_compress,
 )
-from config import get_preset
+from pdft_benchmarks.presets import get_preset
 from data_loading import load_quickdraw
 from evaluation import evaluate_baseline, evaluate_basis_shared
 from generate_report import main as generate_report_main
